1.10.1 Fix Linux and Mac Support
UberDespatch should now build on Linux and Mac, the printing changes
broke support a while back but this has now been corrected but will need
to be tested.

The System.Net.Http package is required, on Linux NuGet can be used to
get this package however the NuGet client must be up to date which means
MonoDevelop 6.1.2+ is required, this cannot be obtained via the apt-get
package and instead must be installed manually from the MonoDevelop
website download.