1.11.0 Solemn Hawthorn
The Manual Carrier now brings up an entry form asking for a Carrier Name
and Tracking Number, these can then be sent to the despatch url if set.
It is recommended to remove edit mode from any manual rules now.

Added a config field for the Despatch URL and PVX Update Template.

Alerts now popup earlier in the processing stage so that rules with edit
mode set to true now show popups before Send is clicked.

Added a DHL Remote Carrier, this is similar to how the Remote (Hive)
Amazon Carrier works where the despatch details are sent to a specified
URL so that a remove system can complete the despatch.

1.10.9

Added a despatch url, if set, the order number, tracking number and
carrier name is sent to this url in json format for external systems to
handle. This is useful if Peoplevox fails to receive additional data
besides the tracking number such as the carrier name.