1.11.1 Linux Printing
The project should once again compile in MonoDevelop under Linux (tested on Ubuntu).
The Pdfium and Codaxy HTML To PDF NuGet packages are both required and NuGet may need to be manually updated for MonoDevelop if not using the latest version.
Printing and html to pdf conversion is now tested and working under Linux, for pdf printing in Linux an 'lp' process is started up instead of using pdfium.